# fix(html): perf and fixes