--- a/crates/swc_html_parser/benches/compare.rs
+++ b/crates/swc_html_parser/benches/compare.rs
@@ -43,7 +43,7 @@ where
             .parse_document_fragment(
                 Element {
                     span: Default::default(),
-                    tag_name: "template".into(),
+                    tag_name: js_word!("template"),
                     namespace: Namespace::HTML,
                     attributes: vec![],
                     is_self_closing: false,