--- a/crates/swc_html_parser/benches/compare.rs
+++ b/crates/swc_html_parser/benches/compare.rs
@@ -1,6 +1,7 @@
 extern crate swc_node_base;
 
 use criterion::{black_box, criterion_group, criterion_main, Bencher, Criterion};
+use swc_atoms::js_word;
 use swc_common::{input::StringInput, FileName, Span, SyntaxContext, DUMMY_SP};
 use swc_html_ast::{Document, DocumentFragment, DocumentMode, Element, Namespace};
 use swc_html_parser::{lexer::Lexer, parser::Parser};