--- a/crates/swc_html_utils/Cargo.toml
+++ b/crates/swc_html_utils/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = [
-    "강동윤 <kdy1997.dev@gmail.com>",
-    "Alexander Akait <sheo13666q@gmail.com>",
+  "강동윤 <kdy1997.dev@gmail.com>",
+  "Alexander Akait <sheo13666q@gmail.com>",
 ]
 description = "Utils for HTML"
 documentation = "https://rustdoc.swc.rs/swc_html_utils/"
@@ -19,4 +19,5 @@ bench = false
 once_cell  = "1.10.0"
 serde      = { version = "1.0.118", features = ["derive"] }
 serde_json = "1.0.61"
+swc_atoms  = { version = "0.4.17", path = "../swc_atoms" }
 swc_common = { version = "0.29.3", path = "../swc_common" }