--- a/crates/swc_html_utils/src/lib.rs
+++ b/crates/swc_html_utils/src/lib.rs
@@ -1,5 +1,6 @@
 use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
+use swc_atoms::JsWord;
 use swc_common::collections::AHashMap;
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -28,21 +29,21 @@ pub struct AttributeInfo {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Element {
-    _extends: Option<Vec<String>>,
+    _extends: Option<Vec<JsWord>>,
     #[serde(flatten)]
-    pub other: AHashMap<String, AttributeInfo>,
+    pub other: AHashMap<JsWord, AttributeInfo>,
 }
 
-pub static HTML_ELEMENTS_AND_ATTRIBUTES: Lazy<AHashMap<String, Element>> = Lazy::new(|| {
-    let default_attributes: AHashMap<String, Element> =
+pub static HTML_ELEMENTS_AND_ATTRIBUTES: Lazy<AHashMap<JsWord, Element>> = Lazy::new(|| {
+    let default_attributes: AHashMap<JsWord, Element> =
         serde_json::from_str(include_str!("../data/html_elements_and_attributes.json"))
             .expect("failed to parse html_elements_and_attributes.json for default attributes");
 
     default_attributes
 });
 
-pub static SVG_ELEMENTS_AND_ATTRIBUTES: Lazy<AHashMap<String, Element>> = Lazy::new(|| {
-    let svg_elements_and_attributes: AHashMap<String, Element> =
+pub static SVG_ELEMENTS_AND_ATTRIBUTES: Lazy<AHashMap<JsWord, Element>> = Lazy::new(|| {
+    let svg_elements_and_attributes: AHashMap<JsWord, Element> =
         serde_json::from_str(include_str!("../data/svg_elements_and_attributes.json"))
             .expect("failed to parse svg_elements_and_attributes.json for default attributes");