<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

Here:
- perf improvement:
```
Gnuplot not found, using plotters backend
Benchmarking html/minify/document/css_spec
Benchmarking html/minify/document/css_spec: Warming up for 3.0000 s
Benchmarking html/minify/document/css_spec: Collecting 10 samples in estimated 5.4698 s (110 iterations)
Benchmarking html/minify/document/css_spec: Analyzing
html/minify/document/css_spec
                        time:   [48.982 ms 49.110 ms 49.259 ms]
                        change: [-12.105% -6.2772% -2.4870%] (p = 0.03 < 0.05)
                        Performance has improved.
```
- bugfix for `EqIgnoreSpan` due `raw`

**BREAKING CHANGE:**

Yes, no we use `JsWord` for utils

**Related issue (if exists):**

No